Split lint-margin files and add a max-lines warning band (#1302)
ManageDialogView.tsx (496 counted lines) and globalConfigThunks.ts (495)
sat within a handful of lines of eslint's max-lines:500 cap, and
registerDatabaseRoutes sat at 49 of funlen's 50-statement limit, with no
signal before any of them tripped.

- Extract the tab-switcher (ManageDialogContent, DirtyAwareTabsTrigger,
  RedeployBanner) out of ManageDialogView.tsx into ManageDialogTabs.tsx:
  496 -> 317 counted lines, new file at 187.
- Extract the cloud provider/context thunk group out of
  globalConfigThunks.ts into globalConfigCloudThunks.ts, re-exported so no
  call site changes: 495 -> 140 counted lines, new file at 358.
- Add a second max-lines rule instance (aliasing the same builtin
  implementation under `lines-budget/approaching-max-lines`) at max:450,
  severity warn, so a file entering the last 50 lines reports it in its
  own lint output. It already surfaces sessionThunks.ts (475),
  EnvironmentDialogView.tsx (469), Titlebar.IdleStatusWidget.tsx (464),
  and types.ts (453) as warnings, none of which are touched here.
- Collapse the 12 repeated repository.NewXRepository(txManager) lines in
  registerDatabaseRoutes into a newDatabaseRepositories(txManager)
  constructor: 49 -> 38 statements.

Closes #1302
